Beggs and Brill correlation

Calculate the Z factor with the Brill-Beggs correlation

z.BeggsBrill(pres.pr, temp.pr, tolerance = 1e-13, verbose = FALSE)

Arguments

pres.pr
pseudo-reduced pressure

temp.pr
pseudo-reduced temperature

tolerance
rounding tolerance to avoid rounding readings that are in the middle of the grid. "tolerance" adds flexibility in deciding point closeness.

verbose
print internal

Examples

## one single z calculation
z.BeggsBrill(pres.pr = 1.5, temp.pr = 2.0)

#> [1] 0.962902

## calculate z for multiple values of Tpr and Ppr
ppr <- c(0.5, 1.5, 2.5, 3.5, 4.5, 5.5, 6.5)
tpr <- c(1.3, 1.5, 1.7, 2)
z.BeggsBrill(pres.pr = ppr, temp.pr = tpr)

#>           0.5       1.5       2.5       3.5       4.5       5.5       6.5
#> 1.3 0.9266436 0.7675523 0.6526911 0.6234648 0.6921991 0.7779095 0.8630653
#> 1.5 0.9555248 0.8618306 0.7945385 0.7691830 0.7828753 0.8248905 0.8837555
#> 1.7 0.9719193 0.9159219 0.8728791 0.8521620 0.8556641 0.8800665 0.9198223
#> 2   0.9853337 0.9629020 0.9471826 0.9404180 0.9443010 0.9593080 0.9848256