Generate a dataset of z values read from Standing-Kats chart

Generate a dataset of z values read from Standing-Kats chart

HY.genDataset7p4t(to_disk = FALSE)

Arguments

to_disk
logical indicator to save Rda file to disk. Default FALSE