Create a tidy table from Ppr and Tpr vectors

Create a tidy table from Ppr and Tpr vectors

createTidyFromMatrix(ppr_vector, tpr_vector, correlation)

Arguments

ppr_vector
a pseudo-reduced pressure vector

tpr_vector
a pseudo-reduced temperature vector

correlation
a z-factor correlation

Examples

ppr <- c(0.5, 1.5, 2.5, 3.5)
tpr <- c(1.05, 1.1, 1.2)
createTidyFromMatrix(ppr, tpr, correlation = "DAK")

#>     Tpr Ppr z.chart    z.calc           dif
#> 1  1.05 0.5   0.829 0.8300683 -0.0010683396
#> 2   1.1 0.5   0.854 0.8570452 -0.0030452477
#> 3   1.2 0.5   0.893 0.8950631 -0.0020631238
#> 4  1.05 1.5   0.253 0.2837318 -0.0307317809
#> 5   1.1 1.5   0.426 0.4463987 -0.0203986679
#> 6   1.2 1.5   0.657 0.6532419  0.0037580632
#> 7  1.05 2.5   0.343 0.3868282 -0.0438282271
#> 8   1.1 2.5   0.393 0.4125200 -0.0195199958
#> 9   1.2 2.5   0.519 0.5180675  0.0009324857
#> 10 1.05 3.5   0.471 0.5063005 -0.0353004935
#> 11  1.1 3.5   0.500 0.5178068 -0.0178068478
#> 12  1.2 3.5   0.565 0.5631805  0.0018194904

createTidyFromMatrix(ppr, tpr, correlation = "BB")

#>     Tpr Ppr z.chart    z.calc           dif
#> 1  1.05 0.5   0.829 0.8325491 -0.0035490638
#> 2   1.1 0.5   0.854 0.8639321 -0.0099321264
#> 3   1.2 0.5   0.893 0.9026461 -0.0096461357
#> 4  1.05 1.5   0.253 0.2851494 -0.0321493974
#> 5   1.1 1.5   0.426 0.4631501 -0.0371500915
#> 6   1.2 1.5   0.657 0.6758659 -0.0188659426
#> 7  1.05 2.5   0.343 0.3333796  0.0096204042
#> 8   1.1 2.5   0.393 0.3848774  0.0081226299
#> 9   1.2 2.5   0.519 0.4977865  0.0212134726
#> 10 1.05 3.5   0.471 0.4569204  0.0140796494
#> 11  1.1 3.5   0.500 0.4993213  0.0006787286
#> 12  1.2 3.5   0.565 0.5605500  0.0044499681