Get a numeric vector of the digitized curves available for Pseudo Reduced Temperature

Get a numeric vector of the digitized curves available for Pseudo Reduced Temperature

getStandingKatzTpr(pprRange = NULL)

Arguments

pprRange
Takes one of 4 values: "lp": low pressure, or "hp" for high pressure; "all": all curves; "common": only curves that are common to hp and lp

Examples

getStandingKatzTpr(pprRange = "lp")

#>  [1] 1.05 1.10 1.20 1.30 1.40 1.50 1.60 1.70 1.80 1.90 2.00 2.20 2.40 2.60 2.80
#> [16] 3.00

getStandingKatzTpr(pprRange = "common")

#>  [1] 1.05 1.10 1.20 1.30 1.40 1.50 1.70 2.00 2.20 2.40 2.60 3.00