Summary statistics table for a correlation

Get error summary statistics for any given compressibility correlation. A quick way to show an error summary between any of the indicated correlations and the Standing-Katz chart.

MSE: Mean Squared Error RMSE: Root Mean Squared Error RSS: Residual sum of Squares RMSLE: Root Mean Squared Logarithmic Error. Penalizes understimation. MAPE: Mean Absolute Percentage Error = AARE MPE: Mean Percentage error = ARE MAE: Mean Absolute Error

z.stats(correlation = "DAK", pprRange = "lp", interval = "coarse")

Arguments

correlation
identifier. Can be "HY", "DAK", "DPR" "N10", "SH"

pprRange
low (lp) or high (hp) chart area of the Standing-Katz chart

interval
quality of the Ppr scale. Coarse: every 1.0; Fine: every 0.5

Examples

# error statistics for the Dranchuk-AbouKassem correlation
z.stats("DAK")

#> # A tibble: 112 x 11
#> # Groups:   Tpr [?]
#>      Tpr   Ppr z.chart    z.calc        RMSE         MPE       MAPE
#>    <chr> <dbl>   <dbl>     <dbl>       <dbl>       <dbl>      <dbl>
#>  1  1.05   0.5   0.829 0.8300683 0.001068340  -0.1288709  0.1288709
#>  2  1.05   1.5   0.253 0.2837318 0.030731781 -12.1469490 12.1469490
#>  3  1.05   2.5   0.343 0.3868282 0.043828227 -12.7779088 12.7779088
#>  4  1.05   3.5   0.471 0.5063005 0.035300493  -7.4947969  7.4947969
#>  5  1.05   4.5   0.598 0.6239783 0.025978253  -4.3441895  4.3441895
#>  6  1.05   5.5   0.727 0.7392097 0.012209748  -1.6794701  1.6794701
#>  7  1.05   6.5   0.846 0.8521762 0.006176173  -0.7300441  0.7300441
#>  8   1.1   0.5   0.854 0.8570452 0.003045248  -0.3565864  0.3565864
#>  9   1.1   1.5   0.426 0.4463987 0.020398668  -4.7884197  4.7884197
#> 10   1.1   2.5   0.393 0.4125200 0.019519996  -4.9669201  4.9669201
#> # ... with 102 more rows, and 4 more variables: MSE <dbl>, RSS <dbl>,
#> #   MAE <dbl>, MAAPE <dbl>

# error statistics for Hall-Yarborough correlation at steps of 0.5 in Ppr
z.stats("HY", interval = "fine")

#> # A tibble: 224 x 11
#> # Groups:   Tpr [?]
#>      Tpr   Ppr z.chart    z.calc        RMSE        MPE      MAPE          MSE
#>    <chr> <dbl>   <dbl>     <dbl>       <dbl>      <dbl>     <dbl>        <dbl>
#>  1  1.05   0.5   0.829 0.8324659 0.003465859  -0.418077  0.418077 1.201218e-05
#>  2  1.05   1.0   0.589 0.6023739 0.013373916  -2.270614  2.270614 1.788616e-04
#>  3  1.05   1.5   0.253 0.3098781 0.056878063 -22.481448 22.481448 3.235114e-03
#>  4  1.05   2.0   0.280 0.3310063 0.051006294 -18.216534 18.216534 2.601642e-03
#>  5  1.05   2.5   0.343 0.3844710 0.041470998 -12.090670 12.090670 1.719844e-03
#>  6  1.05   3.0   0.407 0.4416883 0.034688324  -8.522930  8.522930 1.203280e-03
#>  7  1.05   3.5   0.471 0.4996188 0.028618834  -6.076186  6.076186 8.190377e-04
#>  8  1.05   4.0   0.534 0.5575012 0.023501162  -4.400967  4.400967 5.523046e-04
#>  9  1.05   4.5   0.598 0.6150882 0.017088211  -2.857560  2.857560 2.920070e-04
#> 10  1.05   5.0   0.663 0.6722960 0.009295991  -1.402110  1.402110 8.641545e-05
#> # ... with 214 more rows, and 3 more variables: RSS <dbl>, MAE <dbl>,
#> #   MAAPE <dbl>