Get error summary statistics for any given compressibility correlation

A quick way to show an error summary between any of the indicated correlations and the Standing-Katz chart.

z.stats(correlation = "DAK", pprRange = "lp", interval = "coarse")

Arguments

correlation
identifier. Can be "HY", "DAK", "DPR" "N10", "SH"

pprRange
low (lp) or high (hp) chart area of the Standing-Katz chart

interval
quality of the Ppr scale. Coarse: every 1.0; Fine: every 0.5

Details

MSE: Mean Squared Error RMSE: Root Mean Squared Error RSS: Residual sum of Squares RMSLE: Root Mean Squared Logarithmic Error. Penalizes understimation. MAPE: Mean Absolute Percentage Error = AARE MPE: Mean Percentage error = ARE MAE: Mean Absolute Error

Examples

# error statistics for the Dranchuk-AbouKassem correlation
z.stats("DAK")

#> # A tibble: 112 x 8
#> # Groups:   Tpr [?]
#>      Tpr   Ppr        RMSE        MPE       MAPE          MSE          RSS
#>    <chr> <dbl>       <dbl>      <dbl>      <dbl>        <dbl>        <dbl>
#>  1  1.05   0.5 0.001068340  0.1288709  0.1288709 1.141349e-06 1.141349e-06
#>  2  1.05   1.5 0.030731781 12.1469490 12.1469490 9.444424e-04 9.444424e-04
#>  3  1.05   2.5 0.043828227 12.7779088 12.7779088 1.920913e-03 1.920913e-03
#>  4  1.05   3.5 0.035300493  7.4947969  7.4947969 1.246125e-03 1.246125e-03
#>  5  1.05   4.5 0.025978253  4.3441895  4.3441895 6.748697e-04 6.748697e-04
#>  6  1.05   5.5 0.012209748  1.6794701  1.6794701 1.490779e-04 1.490779e-04
#>  7  1.05   6.5 0.006176173  0.7300441  0.7300441 3.814511e-05 3.814511e-05
#>  8   1.1   0.5 0.003045248  0.3565864  0.3565864 9.273534e-06 9.273534e-06
#>  9   1.1   1.5 0.020398668  4.7884197  4.7884197 4.161057e-04 4.161057e-04
#> 10   1.1   2.5 0.019519996  4.9669201  4.9669201 3.810302e-04 3.810302e-04
#> # ... with 102 more rows, and 1 more variables: MAE <dbl>

# error statistics for Hall-Yarborough correlation at steps of 0.5 in Ppr
z.stats("HY", interval = "fine")

#> # A tibble: 224 x 8
#> # Groups:   Tpr [?]
#>      Tpr   Ppr        RMSE       MPE      MAPE          MSE          RSS
#>    <chr> <dbl>       <dbl>     <dbl>     <dbl>        <dbl>        <dbl>
#>  1  1.05   0.5 0.003465859  0.418077  0.418077 1.201218e-05 1.201218e-05
#>  2  1.05   1.0 0.013373916  2.270614  2.270614 1.788616e-04 1.788616e-04
#>  3  1.05   1.5 0.056878063 22.481448 22.481448 3.235114e-03 3.235114e-03
#>  4  1.05   2.0 0.051006294 18.216534 18.216534 2.601642e-03 2.601642e-03
#>  5  1.05   2.5 0.041470998 12.090670 12.090670 1.719844e-03 1.719844e-03
#>  6  1.05   3.0 0.034688324  8.522930  8.522930 1.203280e-03 1.203280e-03
#>  7  1.05   3.5 0.028618834  6.076186  6.076186 8.190377e-04 8.190377e-04
#>  8  1.05   4.0 0.023501162  4.400967  4.400967 5.523046e-04 5.523046e-04
#>  9  1.05   4.5 0.017088211  2.857560  2.857560 2.920070e-04 2.920070e-04
#> 10  1.05   5.0 0.009295991  1.402110  1.402110 8.641545e-05 8.641545e-05
#> # ... with 214 more rows, and 1 more variables: MAE <dbl>